Add actions registry, event bus class and plugins package
* Add function for setting up all packages and loading modules.
* Break out plate module.
* Move modules to new structure.
* Extract helper for starting and sending commands.
* Update .gitignore.
* Make receive method on Center private.
* Refactor gain module.
* Move gain specific settings to config from command line.
* Require module name in config to setup module.
* Remove bad return from setup_all_modules.
* Move deque of later commands from core to plugins. Multiple
  concurrent plugins can't share the same deque for this purpose.
* Plugins are not allowed to clear the whole event bus, since that
  would interfer with other running plugins. Plugins now have to keep
  track of their own listeners, by storing the function to remove them.
* Add property finished that returns true if no real listeners are
  registered on event bus. Program runs until no listeners are
  registered.
