+++ camacq/data/config.yml
@@ -40,6 +40,11 @@ host: localhost
 last_well: 'U11--V07'
 fields_x: 2
 fields_y: 2
-first_job: 2
-objective: end_63x
-gain_only: false
+
+gain:
+  init_gain: 'path/to/init_gain_file.csv'
+  first_job: 2
+  template_file: 'path/to/template_file.csv'
+  input_gain: 'path/to/input_gain.csv'
+  gain_only: false
+  objective: end_63x  # objective to use for the experiment
